Improved README's usage for custom volumes.
What
===
Added `-v` and `volumes` to the usage.

Why
===
So it's clear that they need to be provided when defining custom volume locations instead of the default `/unison`. The Dockerfile contains a `VOLUME` command which is set to `/unison` when building, which is why the `volumes` definition in the docker-compose file and the `-v` on the command line are unnecessary when it is unchanged.